Task60 Develop get method selectclassesforplan (#65)
closes: #60
* Create Controller for Classes

Created the Controller file for Classes

* Add database seed for Classes

* Add DTO specification for Class objects

* Add entries to Classes in the database

Added some more entries to the table Class so that the software can potentially be tested more easily

* Add GetAvailableClassesForPlan in ClassesController

Added the controller for the class Class and a method GetAvailableClassesForPlan . It returns a list of classes that match the filters and criteria from the use case: "Create Plan".

* Remove reference to TypeItems

Fixed API schemas by replacing the reference to TypeItems with a string.

* Add remaining members in ClassForPlanDTO

Added some remaining members in ClassForPlanDTO in order to satisfy the requirements

* Update data of TypeItems

Updated some FKs in TypeItems to give an item type to some of the classses in the database.